requirements.txt
streamlit
google-generativeai
google-auth